Support np.random.permutation with int argument.
We currently replicate the result of the call to np.random.permutation with
integer argument.  Ideally, this should be parallelized, so that each processor
generates a chunk of the permutation array.  But this is mostly an exercise on
how to call a C++ extension, and we leave the true parallel implementation of
permutation for future work.