Support np.random.permutation with integer argument.
For now, we generate a permutation index in one processor and broadcast it to
all.  Our assumption is that the index is much smaller than the data items we
will shuffle, therefore, the cost of broadcasting an array will be negligible
compared to data shuffling.  We will support fully parallel permutation in the
future.