Fix for Series.head and distribution of StringArrayType as a const_slice
The test 'test_series_head_index_parallel2' from TestSeries was hanging
when executed twice on Windows with MPI run and number of processors equal to 3:
	mpiexec -n 3 python -u -m unittest $TESTS_TO_RUN
	where TESTS_TO_RUN="hpat.tests.TestSeries.test_series_head_parallel1 unittest hpat.tests.TestSeries.test_series_head_parallel1"

The problem was in wrong calculation of n_chars in const_slice_getitem
overload for StringArrayType, which was not correctly reduced by all
processors and had wrong value in some cases. The fix is to simplify the
implementation and to have workers decide whether their part of the array
has to be sent or not basing on the position of their subset in the total
array and the size of requested slice.