Changing tests due to int dtype mismatch on Windows
Reason: difference in dtype of result values occurs due to Numba's rules
for integer typing: http://numba.pydata.org/numba-doc/latest/proposals/integer-typing.html).
Specifically, in modified tests, by the fact that typing of integer constant used as an
argument of binary operation results in producing intp type (int64 on x86_64), which may differ
from the dtype of Series used as another argument (on Linux it's int64 and on Windows int32).
At the same time NumPy operates differently, so that the dtype of resulting series in expressions
like "S1 + 42" is the same as the dtype of S1.