Change style of pd.Series overload and fix check of None index
Reason: this solves a problem of calling pd.Series(data, index)
where index argument is not omitted and has a Numba's type
numba.types.NoneType, in which case the old check doesn't work
and we attempt to call fix_df_array function for none.

Also changes function design to use Numba's branch pruning
and have one implementation in the overload instead of two.